+++ CHANGELOG.md
@@ -6,25 +6,46 @@ SPDX-License-Identifier: MIT
 
 # Changelog
 
-## [TOLP-10145]("https://jira.sanger.ac.uk/browse/TOLP-10145")
+## [TOLP-10027](https://jira/sanger.ac.uk/browse/TOLP-10027)
+
+17-08-2026
+
+- `tol-ui 4.0.13`
+
+## [TOLP-10094](https://jira.sanger.ac.uk/browse/TOLP-9958)
+
+13-08-2026
+
+- Removed old `topup` role
+- Added new `core lab top up` and `core lab sample export` roles
+- Associated roles and actions based on supplied list
+- Re-assigned `topup` users to appropriate new roles.
+
+## [TOLP-10120](https://jira.sanger.ac.uk/browse/TOLP-10120)
+
+13-08-2026
+
+- `tol-ui 4.0.12`
+
+## [TOLP-10145](https://jira.sanger.ac.uk/browse/TOLP-10145)
 
 27-07-2026
 
 - Fixed bug: defaultFilter acting as a base stopping live filters being removed
 
-## [TOLP-10035]("https://jira.sanger.ac.uk/browse/TOLP-10035")
+## [TOLP-10035](https://jira.sanger.ac.uk/browse/TOLP-10035)
 
 27-07-2026
 
 - Advanced translators and the ability to exclude incoming filters
 
-## [TOLP-10073]("https://jira.sanger.ac.uk/browse/TOLP-10073")
+## [TOLP-10073](https://jira.sanger.ac.uk/browse/TOLP-10073)
 
 27-07-2026
 
 - `tol-ui 4.0.5`
 
-## [TOLP-9742]("https://jira.sanger.ac.uk/browse/TOLP-9742")
+## [TOLP-9742](https://jira.sanger.ac.uk/browse/TOLP-9742)
 
 22-04-2026